Transformed instruction from struct to enum (#4)
* Transformed instruction from struct to enum

* Added instruction unit tests

* Fix some error/warnings

* Address clippy errors

* remove Operand enum

* refactor instructions

* Use read_imm8()

* Fix formatting

+++ src/main.rs
@@ -5,10 +5,10 @@ use std::fs;
 
 fn main() -> Result<(), std::io::Error> {
     let buffer = fs::read("../roms/invaders.h")?;
-    
+
     let mut pc = 0;
     while pc < buffer.len() {
-        let instr = Instruction::from(&buffer[usize::from(pc)..]);
+        let instr = Instruction::from(&buffer[pc..]);
         println!("{:?}", instr);
         pc += instr.size() as usize;
     }
